% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noswimming.R
\name{lcp}
\alias{lcp}
\title{Create least cost path between two points}
\usage{
lcp(x, surf = NULL)
}
\arguments{
\item{x}{x  matrix of 2 points, lon,lat}

\item{surf}{surface for use by create_cs}
}
\value{
terra vect path
}
\description{
By default we assuming swimming in the ocean, swimming on land is not allowed.
}
\examples{
lcp(cbind(c(-90, 90), c(-70, -60)))
}
