% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noswimming.R
\name{make_global_topo}
\alias{make_global_topo}
\title{Make a global topography in a projection, with input longlat points in the centre}
\usage{
make_global_topo(x, ncols = 1024, nrows = ncols)
}
\arguments{
\item{x}{matrix of 2 points, lon,lat}

\item{ncols}{number of columns of the output (default is 1024, which is about 26km resolution for the globe)}

\item{nrows}{ditto ncols (default is ncols)}
}
\value{
raster of global topography
}
\description{
Input is two longlat points
}
\details{
CRS is Lambert Azimuthal Equal Area, determined around the centre of the input points.
}
\examples{
make_global_topo(cbind(c(-90, 90), c(-70, -60)))
}
