% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noswimming.R
\name{make_na_surface}
\alias{make_na_surface}
\title{Make surface for cost surface creation}
\usage{
make_na_surface(x, maxvalue = 0, minvalue = -Inf, fillvalue = 1)
}
\arguments{
\item{x}{a raster, can have missing values}

\item{maxvalue}{maximum value for the reachable areas (default to 0)}

\item{minvalue}{minimum value for the reachable areas (default to -Inf)}

\item{fillvalue}{1 (should be positive)}
}
\value{
a raster
}
\description{
This sets a DEM to have NA values for anywhere out of bounds, with a fill value of 1
everywhere else.
}
\details{
If this isn't a good input to create_cs create a suitable one, perhaps starting with make_global_topo().

Note that fill must be positive value else it seems to fail with leastcostpath.
}
\examples{
make_na_surface(terra::rast(volcano), maxvalue = 130)
}
